chore: consolidate type definitions
