--- apps/campfire/globals.d.ts
@@ -1,5 +1,0 @@
-declare global {
-  interface Window {
-    storyFormat: (input: string) => string
-  }
-}
