--- matchers.d.ts
@@ -1,8 +1,0 @@
-import { TestingLibraryMatchers } from '@testing-library/jest-dom/matchers'
-import { Matchers, AsymmetricMatchers } from 'bun:test'
-
-declare module 'bun:test' {
-  interface Matchers<T>
-    extends TestingLibraryMatchers<typeof expect.stringContaining, T> {}
-  interface AsymmetricMatchers extends TestingLibraryMatchers {}
-}
